rpk cluster brokers: new subcommand group
Move decommission, decommission-status, and
recommission from `rpk redpanda admin brokers`
to a new `rpk cluster brokers` group.
Cluster-level operations belong under
`rpk cluster`; the admin tree overlaps
conceptually with cluster commands and is
being deprecated.

The old commands remain functional as thin
shims that delegate to the new package, are
hidden from help, and print a deprecation
notice pointing at the new location.
