+++ src/go/rpk/pkg/cli/cluster/brokers/BUILD
@@ -1,20 +1,17 @@
-load("@rules_go//go:def.bzl", "go_library")
+load("@rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "brokers",
     srcs = [
         "brokers.go",
         "decommission.go",
-        "decommission-status.go",
-        "list.go",
+        "decommission_status.go",
         "recommission.go",
     ],
-    importpath = "github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/redpanda/admin/brokers",
+    importpath = "github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cluster/brokers",
     visibility = ["//visibility:public"],
     deps = [
         "//src/go/rpk/pkg/adminapi",
-        "//src/go/rpk/pkg/cli/cluster/brokers",
-        "//src/go/rpk/pkg/cobraext",
         "//src/go/rpk/pkg/config",
         "//src/go/rpk/pkg/out",
         "//src/go/rpk/pkg/redpanda",
@@ -22,5 +19,18 @@ go_library(
         "@com_github_redpanda_data_common_go_rpadmin//:rpadmin",
         "@com_github_spf13_afero//:afero",
         "@com_github_spf13_cobra//:cobra",
+        "@com_github_twmb_types//:types",
+    ],
+)
+
+go_test(
+    name = "brokers_test",
+    srcs = ["decommission_status_test.go"],
+    embed = [":brokers"],
+    deps = [
+        "//src/go/rpk/pkg/config",
+        "//src/go/rpk/pkg/out",
+        "@com_github_redpanda_data_common_go_rpadmin//:rpadmin",
+        "@com_github_stretchr_testify//require",
     ],
 )
